This PR adds the following:
* IDA brancher. Specify `--brancher=ida` to use IDA's information of
branch points / jump edges.

* More IDA loader functionality. The `--loader=ida` option will now also
add IDA's "synthetic" sections to the binary image, such as `extern`.

* Linux kernel module symbols. Kernel symbols are fixed up at runtime.
To get the symbols statically in BAP, we need a) relocation information
of which symbols are called at which addresses, and b) to fix up the CFG
at the callsites to kernel symbols. Part (b) is performed by the
`ida_ko_symbol_mapper` plugin.